@@ -1,35 +1,35 @@
 /*
- * Copyright 2016 Simon Edwards <simon@simonzone.com>
+ * Copyright 2016-2017 Simon Edwards <simon@simonzone.com>
  *
  * This source code is licensed under the MIT license which is detailed in the LICENSE.txt file.
  */
-import * as CommandPaletteTypes from './gui/CommandPaletteTypes';
+import * as CommandPaletteFunctons from './CommandPaletteFunctions';
 
-export interface Commandable {
-  executeCommand(commandId: string): void;
+export interface CommandExecutor {
+  executeCommand(commandId: string, commandArguments?: object): void;
+}
+
+export interface Commandable extends CommandExecutor {
+  getCommandPaletteEntries(commandableStack: Commandable[]): CommandEntry[];
 }
 
 export function isCommandable(instance: any): instance is Commandable {
   if (instance === null || instance === undefined) {
     return false;
   }
-  return (<Commandable> instance).executeCommand !== undefined;
+  return (<Commandable> instance).executeCommand !== undefined && (<Commandable> instance).getCommandPaletteEntries !== undefined;
 }
 
-export interface CommandEntry extends CommandPaletteTypes.CommandEntry {
-  id: string;
-  iconLeft?: string;
-  iconRight?: string;
-  label: string;
-  shortcut?: string;
-  target: Commandable;
+export function dispatchCommandPaletteRequest(element: Commandable & HTMLElement): void {
+  const commandPaletteRequestEvent = new CustomEvent(EVENT_COMMAND_PALETTE_REQUEST, {bubbles: true, composed: true});
+  commandPaletteRequestEvent.initCustomEvent(EVENT_COMMAND_PALETTE_REQUEST, true, true, null);
+  element.dispatchEvent(commandPaletteRequestEvent);
+}
+
+export interface CommandEntry extends CommandPaletteFunctons.CommandMenuItem {
+  commandExecutor: CommandExecutor;
+  commandArguments?: object;
 }
 
 export const EVENT_COMMAND_PALETTE_REQUEST = "EVENT_COMMAND_PALETTE_REQUEST";
 export const COMMAND_OPEN_COMMAND_PALETTE = "openCommandPalette";
-
-export interface CommandPaletteRequest {
-  commandEntries: CommandEntry[];
-  srcElement: HTMLElement;
-  contextElement: HTMLElement;
-}
