@@ -6,13 +6,11 @@
 
 import * as fs from 'fs';
 import * as Util from './gui/Util';
-import * as VirtualScrollArea from './VirtualScrollArea';
+import * as CodeMirrorOperation from './CodeMirrorOperation';
+import {VirtualScrollable, SetterState} from './VirtualScrollArea';
 import {Mode, VisualState, CursorMoveDetail} from './ViewerElementTypes';
 import {ThemeableElementBase} from './ThemeableElementBase';
-import * as CodeMirrorOperation from './CodeMirrorOperation';
 
-type VirtualScrollable = VirtualScrollArea.VirtualScrollable;
-type SetterState = VirtualScrollArea.SetterState;
 
 export abstract class ViewerElement extends ThemeableElementBase implements VirtualScrollable {
   
